autotrain/requirements.txt
python-dotenv
datasets
wandb
huggingface_hub
autotrain-advanced